Move PAK merging into main mod
