--- Extension/p5r.modloader.pak/CHANGELOG.MD
@@ -1,2 +1,0 @@
-#1.0.0
-Initial Commit
