--- Extension/p5r.modloader.pak/ModConfig.json
@@ -1,96 +1,0 @@
-{
-  "ModId": "p5rpc.modloader.pak",
-  "ModName": "PAK Emulator Support for Persona Essentials",
-  "ModAuthor": "ltsophia",
-  "ModVersion": "1.0.0",
-  "ModDescription": "Mod which provides base modding support as well as miscellaneous basic quality of life fixes. See readme for Special Thanks!",
-  "ModDll": "p5rpc.modloader.pak.dll",
-  "ModIcon": "Preview.png",
-  "ModR2RManagedDll32": "x86/p5rpc.modloader.pak.dll",
-  "ModR2RManagedDll64": "x64/p5rpc.modloader.pak.dll",
-  "ModNativeDll32": "",
-  "ModNativeDll64": "",
-  "Tags": [],
-  "CanUnload": false,
-  "HasExports": false,
-  "IsLibrary": true,
-  "ReleaseMetadataFileName": "p5rpc.modloader.pak.ReleaseMetadata.json",
-  "PluginData": {
-    "GitHubDependencies": {
-      "IdToConfigMap": {
-        "reloaded.sharedlib.hooks": {
-          "Config": {
-            "UserName": "Reloaded.SharedLib.Hooks.ReloadedII",
-            "RepositoryName": "Sewer56",
-            "UseReleaseTag": true,
-            "AssetFileName": "reloaded.sharedlib.hooks.zip"
-          },
-          "ReleaseMetadataName": "Sewer56.Update.ReleaseMetadata.json"
-        },
-        "Reloaded.Memory.SigScan.ReloadedII": {
-          "Config": {
-            "UserName": "Reloaded-Project",
-            "RepositoryName": "Reloaded.Memory.SigScan",
-            "UseReleaseTag": false,
-            "AssetFileName": "Mod.zip"
-          },
-          "ReleaseMetadataName": "Reloaded.Memory.SigScan.ReloadedII.ReleaseMetadata.json"
-        },
-        "crifs.v2.hook": {
-          "Config": {
-            "UserName": "Sewer56",
-            "RepositoryName": "CriFs.V2.Hook.ReloadedII",
-            "UseReleaseTag": false,
-            "AssetFileName": "Mod.zip"
-          },
-          "ReleaseMetadataName": "CriFs.V2.Hook.ReleaseMetadata.json"
-        },
-        "reloaded.universal.fileemulationframework.pak": {
-          "Config": {
-            "UserName": "Sewer56",
-            "RepositoryName": "FileEmulationFramework",
-            "UseReleaseTag": false,
-            "AssetFileName": "Mod.zip"
-          },
-          "ReleaseMetadataName": "PAK.Stream.Emulator.ReleaseMetadata.json"
-        },
-        "reloaded.universal.fileemulationframework": {
-          "Config": {
-            "UserName": "Sewer56",
-            "RepositoryName": "FileEmulationFramework",
-            "UseReleaseTag": false,
-            "AssetFileName": "Mod.zip"
-          },
-          "ReleaseMetadataName": "Reloaded.Universal.FileEmulationFramework.ReleaseMetadata.json"
-        }
-      }
-    },
-    "NuGet": {
-      "AllowUpdateFromAnyRepository": false,
-      "DefaultRepositoryUrls": [
-        "http://packages.sewer56.moe:5000/v3/index.json"
-      ]
-    },
-    "GitHubRelease": {
-      "UserName": "Sewer56",
-      "RepositoryName": "p5rpc.modloader.pak",
-      "UseReleaseTag": true,
-      "AssetFileName": "Mod.zip"
-    }
-  },
-  "IsUniversalMod": false,
-  "ModDependencies": [
-    "Reloaded.Memory.SigScan.ReloadedII",
-    "reloaded.sharedlib.hooks",
-    "crifs.v2.hook",
-    "reloaded.universal.fileemulationframework.pak",
-    "p5rpc.modloader"
-  ],
-  "OptionalDependencies": [],
-  "SupportedAppId": [
-    "p3p.exe",
-    "p4g.exe",
-    "p5r.exe"
-  ],
-  "ProjectUrl": "https://sewer56.dev/p5rpc.modloader/"
-}
