--- Extension/p5r.modloader.pak/Template/Startup.cs
@@ -1,92 +1,0 @@
-﻿/*
- * This file and other files in the `Template` folder are intended to be left unedited (if possible),
- * to make it easier to upgrade to newer versions of the template.
-*/
-
-using p5r.modloader.pak.Template.Configuration;
-using Reloaded.Mod.Interfaces;
-using Reloaded.Mod.Interfaces.Internal;
-
-namespace p5r.modloader.pak.Template;
-
-public class Startup : IMod
-{
-    /// <summary>
-    /// Used for writing text to the Reloaded log.
-    /// </summary>
-    private ILogger _logger = null!;
-
-    /// <summary>
-    /// Provides access to the mod loader API.
-    /// </summary>
-    private IModLoader _modLoader = null!;
-
-    /// <summary>
-    /// Stores the contents of your mod's configuration. Automatically updated by template.
-    /// </summary>
-    private Config _configuration = null!;
-
-    /// <summary>
-    /// Configuration of the current mod.
-    /// </summary>
-    private IModConfig _modConfig = null!;
-
-    /// <summary>
-    /// Encapsulates your mod logic.
-    /// </summary>
-    private ModBase _mod = new Mod();
-
-    /// <summary>
-    /// Entry point for your mod.
-    /// </summary>
-    public void StartEx(IModLoaderV1 loaderApi, IModConfigV1 modConfig)
-    {
-        _modLoader = (IModLoader)loaderApi;
-        _modConfig = (IModConfig)modConfig;
-        _logger = (ILogger)_modLoader.GetLogger();
-
-        // Your config file is in Config.json.
-        // Need a different name, format or more configurations? Modify the `Configurator`.
-        // If you do not want a config, remove Configuration folder and Config class.
-        var configurator = new Configurator(_modLoader.GetModConfigDirectory(_modConfig.ModId));
-        _configuration = configurator.GetConfiguration<Config>(0);
-        _configuration.ConfigurationUpdated += OnConfigurationUpdated;
-
-        // Please put your mod code in the class below,
-        // use this class for only interfacing with mod loader.
-        _mod = new Mod(new ModContext()
-        {
-            Logger = _logger,
-            ModLoader = _modLoader,
-            ModConfig = _modConfig,
-            Owner = this,
-            Configuration = _configuration,
-        });
-    }
-
-    private void OnConfigurationUpdated(IConfigurable obj)
-    {
-        /*
-            This is executed when the configuration file gets 
-            updated by the user at runtime.
-        */
-
-        // Replace configuration with new.
-        _configuration = (Config)obj;
-        _mod.ConfigurationUpdated(_configuration);
-    }
-
-    /* Mod loader actions. */
-    public void Suspend() => _mod.Suspend();
-    public void Resume() => _mod.Resume();
-    public void Unload() => _mod.Unload();
-
-    /*  If CanSuspend == false, suspend and resume button are disabled in Launcher and Suspend()/Resume() will never be called.
-        If CanUnload == false, unload button is disabled in Launcher and Unload() will never be called.
-    */
-    public bool CanUnload() => _mod.CanUnload();
-    public bool CanSuspend() => _mod.CanSuspend();
-
-    /* Automatically called by the mod loader when the mod is about to be unloaded. */
-    public Action Disposing => () => _mod.Disposing();
-}
