--- Extension/p5r.modloader.pak/p5rpc.modloader.pak.csproj
@@ -1,55 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0-windows</TargetFramework>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-    <LangVersion>preview</LangVersion>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <OutputPath>$(RELOADEDIIMODS)/p5rpc.modloader.pak</OutputPath>
-    <Deterministic>true</Deterministic>
-    <ImplicitUsings>enable</ImplicitUsings>
-
-    <!-- IL Trimming Support for Framework Dependent Apps. Unofficial.
-         If you set this to true, trimming will be enabled when you publish the mod.
-         Test trimming with: `BuildLinked.ps1` in PowerShell. -->
-    <ReloadedILLink>true</ReloadedILLink>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ReloadedLinkRoots Include="p5r.modloader.pak" />
-    <!-- Add assemblies to be trimmed. You might need to add their dependencies too!  -->
-    <!-- <ReloadedLinkAssemblies Include="Reloaded.Memory" /> -->
-  </ItemGroup>
-
-  <Import Project="Reloaded.Trimming.targets" />
-
-  <ItemGroup>
-    <None Remove="ModConfig.json" />
-    <None Remove="BuildLinked.ps1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="ModConfig.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-    <None Update="Preview.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Reloaded.Memory" Version="8.0.1" />
-    <PackageReference Include="Reloaded.Memory.SigScan.ReloadedII.Interfaces" Version="1.2.0" />
-    <PackageReference Include="Reloaded.Mod.Interfaces" Version="2.3.0" ExcludeAssets="runtime" />
-    <PackageReference Include="Reloaded.SharedLib.Hooks" Version="1.9.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\p5rpc.modloader\p5rpc.modloader.csproj" />
-    <ProjectReference Include="..\..\Persona.Merger.Common\Persona.Merger.Common.csproj" />
-    <ProjectReference Include="..\..\Submodules\CriFs.V2.Hook.ReloadedII\Interfaces\CriFs.V2.Hook.Interfaces\CriFs.V2.Hook.Interfaces.csproj" />
-  </ItemGroup>
-
-</Project>
