+++ p5rpc.modloader/Mod_Merging.cs
@@ -28,7 +28,8 @@ public partial class Mod
         var mergeUtils = new MergeUtils(_criFsApi);
         List<IFileMerger> fileMergers = new()
         {
-            new TblMerger(mergeUtils, _logger, _mergedFileCache, _criFsApi, Game)
+            new TblMerger(mergeUtils, _logger, _mergedFileCache, _criFsApi, Game),
+            new PakMerger(mergeUtils, _logger, _mergedFileCache, _criFsApi, _pakEmulator)
         };
         
         foreach (var fileMerger in fileMergers)
