+++ p5rpc.modloader/p5rpc.modloader.csproj
@@ -51,6 +51,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Persona.Merger.Common\Persona.Merger.Common.csproj" />
     <ProjectReference Include="..\Submodules\CriFs.V2.Hook.ReloadedII\Interfaces\CriFs.V2.Hook.Interfaces\CriFs.V2.Hook.Interfaces.csproj" />
+    <ProjectReference Include="..\Submodules\FileEmulationFramework\Emulator\Interfaces\PAK.Stream.Emulator.Interfaces\PAK.Stream.Emulator.Interfaces.csproj" />
   </ItemGroup>
 
 </Project>
